Document `auth-source-pass-get` alternative to get secrets from pass
It's available from Emacs 26 and autoloaded from Emacs 27.

Auth-sources works with pass if you add `password-store` to `auth-sources`, but
`auth-source-pass-get` should work without that, out-of-the-box. Its interface
is a bit different than using `auth-source-pick-first-password` with the
`password-store` backend, as it wants the entire path to the password like
`pass`, instead of searching through hostnames via matching rules like
auth-sources does.

Still, I only put it into a comment as an alternative next to the
`process-lines` method. Because strictly speaking it's in the heading how to do
things without `auth-sources`. It doesn't rely on the list-variable
`auth-sources` but still loads the `auth-sources.el` library.