Roll own csv parsing, remove `pcsv` dependency
Parse the CSV retrieved from Awesome Prompts [1] on our own. This
removes the dependency on the pcsv package which is not standard.

While there is certainly an argument to be made against NIH, CSV parsing
should also be do-able with reasonable added complexity, at least if the
format is somewhat well-known. A dependency on a third party package
just for this seems unnecessary.

This was added in response to issue #126. I am using this parser
suggested here in claude-shell [2] since quite a while and it worked
well enough - and as far as I can tell also correct.

[1] https://raw.githubusercontent.com/f/awesome-chatgpt-prompts/main/prompts.csv
[2] https://github.com/arminfriedl/claude-shell