fix: Foreign buffer fontification (fixes #375)
Fix fontification of arbitrary buffers. Currently, with streaming
enabled, it is possible to switch away from the chatgpt-shell buffer
while it is finishing writing the response. Once this happened the
`on-command-finished` puts markdown overlays in the current buffer,
which might not be the chatgpt-shell buffer.

Wrap markdown overlay fontification in
`(with-current-buffer (chatgpt-shell--primary-buffer))` s.t. it puts
the overlays in the `chatgpt-shell` buffer.